+++ examples/react-dev-demo/src/components/AppContainer/components/DemoUseFlagship/components/qa/PlayWithHits/index.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { useFlagship } from '@flagship.io/react-sdk';
-import CodeBlock from '../../../../../../common/CodeBlock';
 import { Button, Form, Col, Nav } from 'react-bootstrap';
 import { Formik } from 'formik';
 import JSONInput from 'react-json-editor-ajrm';
 import locale from 'react-json-editor-ajrm/locale/en';
-import flagship from '@flagship.io/js-sdk';
+import { HitShape } from '@flagship.io/js-sdk';
 
 const PlayWithHits: React.FC = () => {
     const defaultTransactionHitPayload = {
@@ -21,16 +20,14 @@ const PlayWithHits: React.FC = () => {
             paymentMethod: 'yoloPaymentMethod',
             itemCount: 2,
             couponCode: 'YOLOCOUPON',
-            documentLocation:
-                'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
+            documentLocation: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
             pageTitle: 'YoloScreen'
         }
     };
     const defaultScreenHitPayload = {
         type: 'Screen',
         data: {
-            documentLocation:
-                'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
+            documentLocation: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
             pageTitle: 'YoloScreen'
         }
     };
@@ -41,8 +38,7 @@ const PlayWithHits: React.FC = () => {
             action: 'signOff',
             label: 'yolo label ;)',
             value: 123,
-            documentLocation:
-                'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
+            documentLocation: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
             pageTitle: 'YoloTitle'
         }
     };
@@ -55,8 +51,7 @@ const PlayWithHits: React.FC = () => {
             code: 'yoloCode',
             category: 'yoloCategory',
             quantity: 1234444,
-            documentLocation:
-                'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
+            documentLocation: 'http%3A%2F%2Fabtastylab.com%2F60511af14f5e48764b83d36ddb8ece5a%2F',
             pageTitle: 'YoloScreen'
         }
     };
@@ -68,12 +63,8 @@ const PlayWithHits: React.FC = () => {
     };
     const [hasError, setError] = React.useState(false);
     const [hasError2, setError2] = React.useState(false);
-    const [currentTestedHit, setHitToTest] = React.useState<string>(
-        'transaction'
-    );
-    const [hitPayload, setHitPayload] = React.useState(
-        payloads[currentTestedHit]
-    );
+    const [currentTestedHit, setHitToTest] = React.useState<string>('transaction');
+    const [hitPayload, setHitPayload] = React.useState(payloads[currentTestedHit]);
 
     const [hitsPayload, setHitsPayload] = React.useState([
         { ...payloads.transaction },
@@ -81,43 +72,32 @@ const PlayWithHits: React.FC = () => {
         { ...payloads.event },
         { ...payloads.item }
     ]);
-    React.useEffect(() => setHitPayload(payloads[currentTestedHit]), [
-        currentTestedHit
-    ]);
+    const update = () => {
+        setHitPayload(payloads[currentTestedHit]);
+    };
+    React.useEffect(update, [currentTestedHit]);
     const { hit: fsHit } = useFlagship();
     return (
         <>
             <p>Send a hit, demo: </p>
             <Nav variant="tabs" defaultActiveKey="hitTransaction">
                 <Nav.Item>
-                    <Nav.Link
-                        eventKey="hitTransaction"
-                        onClick={() => setHitToTest('transaction')}
-                    >
+                    <Nav.Link eventKey="hitTransaction" onClick={() => setHitToTest('transaction')}>
                         Transaction Hit
                     </Nav.Link>
                 </Nav.Item>
                 <Nav.Item>
-                    <Nav.Link
-                        eventKey="hitScreen"
-                        onClick={() => setHitToTest('screen')}
-                    >
+                    <Nav.Link eventKey="hitScreen" onClick={() => setHitToTest('screen')}>
                         Screen Hit
                     </Nav.Link>
                 </Nav.Item>
                 <Nav.Item>
-                    <Nav.Link
-                        eventKey="hitEvent"
-                        onClick={() => setHitToTest('event')}
-                    >
+                    <Nav.Link eventKey="hitEvent" onClick={() => setHitToTest('event')}>
                         Event Hit
                     </Nav.Link>
                 </Nav.Item>
                 <Nav.Item>
-                    <Nav.Link
-                        eventKey="hitItem"
-                        onClick={() => setHitToTest('item')}
-                    >
+                    <Nav.Link eventKey="hitItem" onClick={() => setHitToTest('item')}>
                         Item Hit
                     </Nav.Link>
                 </Nav.Item>
@@ -140,22 +120,12 @@ const PlayWithHits: React.FC = () => {
                     fsHit.send({ ...values.hitPayload });
                 }}
             >
-                {({
-                    handleSubmit,
-                    handleChange,
-                    handleBlur,
-                    setFieldValue,
-                    values,
-                    touched,
-                    isValid,
-                    errors
-                }) => (
+                {({ handleSubmit, handleChange, handleBlur, setFieldValue, values, touched, isValid, errors }) => (
                     <Form noValidate onSubmit={handleSubmit}>
                         <Form.Group as={Col} md="12" controlId="settingsForm">
-                            <Form.Label>
-                                {currentTestedHit} hit payload
-                            </Form.Label>
+                            <Form.Label>{currentTestedHit} hit payload</Form.Label>
                             <JSONInput
+                                waitAfterKeyPress={3000}
                                 id="hitPayload"
                                 placeholder={values.hitPayload}
                                 locale={locale}
@@ -163,11 +133,7 @@ const PlayWithHits: React.FC = () => {
                                 width="100%"
                                 onChange={({ error, jsObject }) => {
                                     if (!error) {
-                                        setFieldValue(
-                                            'hitPayload',
-                                            jsObject || {},
-                                            true
-                                        );
+                                        setFieldValue('hitPayload', jsObject || {}, true);
                                         setError(false);
                                     } else {
                                         setError(true);
@@ -195,10 +161,7 @@ const PlayWithHits: React.FC = () => {
                     </Form>
                 )}
             </Formik>
-            <div className="mv3">
-                Have a look to logs + network on your browser inspect tools to
-                see results.
-            </div>
+            <div className="mv3">Have a look to logs + network on your browser inspect tools to see results.</div>
             <Formik
                 initialValues={{
                     hitsPayload
@@ -213,25 +176,15 @@ const PlayWithHits: React.FC = () => {
                 onSubmit={(values, { setSubmitting }) => {
                     setSubmitting(false);
                     setHitsPayload([...values.hitsPayload]);
-                    fsHit.sendMultiple([
-                        ...values.hitsPayload
-                    ] as flagship.HitShape[]);
+                    fsHit.sendMultiple([...values.hitsPayload] as HitShape[]);
                 }}
             >
-                {({
-                    handleSubmit,
-                    handleChange,
-                    handleBlur,
-                    setFieldValue,
-                    values,
-                    touched,
-                    isValid,
-                    errors
-                }) => (
+                {({ handleSubmit, handleChange, handleBlur, setFieldValue, values, touched, isValid, errors }) => (
                     <Form noValidate onSubmit={handleSubmit}>
                         <Form.Group as={Col} md="12" controlId="settingsForm">
                             <Form.Label>sendMultiple argument</Form.Label>
                             <JSONInput
+                                waitAfterKeyPress={3000}
                                 id="hitsPayload"
                                 placeholder={values.hitsPayload}
                                 locale={locale}
@@ -239,11 +192,7 @@ const PlayWithHits: React.FC = () => {
                                 width="100%"
                                 onChange={({ error, jsObject }) => {
                                     if (!error) {
-                                        setFieldValue(
-                                            'hitsPayload',
-                                            jsObject || {},
-                                            true
-                                        );
+                                        setFieldValue('hitsPayload', jsObject || {}, true);
                                         setError2(false);
                                     } else {
                                         setError2(true);
@@ -262,9 +211,7 @@ const PlayWithHits: React.FC = () => {
                                 variant="secondary"
                                 type="submit"
                                 style={{
-                                    cursor: hasError2
-                                        ? 'not-allowed'
-                                        : 'pointer'
+                                    cursor: hasError2 ? 'not-allowed' : 'pointer'
                                 }}
                             >
                                 Send multiple hits