+++ examples/react-dev-demo/src/components/QaHeader/index.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import { Navbar, Nav, Form, NavDropdown } from 'react-bootstrap';
-import Logo from '../../assets/Flagship-horizontal-product-white.png';
+import { Form, Nav, Navbar } from 'react-bootstrap';
+
 import { AppSettings, SettingContext } from '../../App';
 import SettingsModal from './components/SettingsModal';
 
@@ -29,15 +29,11 @@ const QaHeader: React.FC = () => {
                 </Nav>
 
                 <Form inline>
-                    <Nav.Link onClick={() => setQA({ ...QA, enabled: false })}>
-                        Switch off
-                    </Nav.Link>
+                    <Nav.Link onClick={() => setQA({ ...QA, enabled: false })}>Switch off</Nav.Link>
                 </Form>
             </Navbar>
             <SettingsModal
-                onHide={() =>
-                    setQA({ ...QA, show: { ...QA.show, settingsModal: false } })
-                }
+                onHide={() => setQA({ ...QA, show: { ...QA.show, settingsModal: false } })}
                 show={QA.show.settingsModal}
             ></SettingsModal>
         </>