+++ examples/react-dev-demo/src/config/index.tsx
@@ -1,8 +1,8 @@
-import { FlagshipReactSdkConfig } from '../../../../dist';
-
 const config = {
     envId: 'bn1ab7m56qolupi5sa0g',
     fetchNow: true,
+    decisionMode: 'API',
+    pollingInterval: 5,
     enableConsoleLogs: true,
     enableErrorLayout: true,
     nodeEnv: 'production',
@@ -19,6 +19,8 @@ const config = {
         },
         envId: ['bn1ab7m56qolupi5sa0g_fake_1', 'bn1ab7m56qolupi5sa0g_fake_2'],
         nodeEnv: ['production', 'development'],
+        pollingInterval: [1, 2, 5, 10],
+        decisionMode: ['API', 'Bucketing'],
         visitorId: ['test-vid', 'test-vid_2', 'test-vid_3']
     }
 };