+++ examples/react-ecommerce-demo/components/header.js
@@ -1,14 +1,14 @@
-import dynamic from "next/dynamic";
-import { view } from "@risingstack/react-easy-state";
+import dynamic from 'next/dynamic';
+import { view } from '@risingstack/react-easy-state';
 
-const FSWrapper = dynamic(() => import("./fsWrapper"), { ssr: false });
+const FSWrapper = dynamic(() => import('./fsWrapper'), { ssr: false });
 
-import HeaderInner from "./headerInner";
+import HeaderInner from './headerInner';
 
-export default view(({ cartObj, fsFlags, fsInfos }) => {
-  return (
-    <FSWrapper fsInfos={fsInfos}>
-      <HeaderInner cartObj={cartObj} />
-    </FSWrapper>
-  );
+export default view(({ cartObj, fsInfos }) => {
+    return (
+        <FSWrapper fsInfos={fsInfos}>
+            <HeaderInner cartObj={cartObj} />
+        </FSWrapper>
+    );
 });